+++ CHANGELOG.md
@@ -8,6 +8,12 @@
   - `changes/2` allows to fetch the changes of an invidual source record
 * Update Postgrex to 0.15.11 and drop local `Xid8` type
 
+### Changed
+
+* Moved top-level functions to nested modules `Transaction` and `Multi`
+* Made `table_pk` be `NULL` when `primary_key_columns` is an empty array
+* Default `primary_key_columns` to `["id"]`
+
 ## [0.1.0] - 2021-09-01
 
 * Initial release.